% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_housing.R
\docType{data}
\name{calHousing}
\alias{calHousing}
\title{California Median House Prices}
\format{An object of class \code{"data.frame"}.}
\usage{
data(calHousing)
}
\description{
Data of California median housing prices
}
\examples{
data(calHousing)
medianValue <- log(calHousing$medianValue)
x.matrix    <- data.matrix(calHousing[,-match("medianValue", colnames(calHousing))])

# Fit with oem

calfit    <- oem(x = x.matrix, y = medianValue, 
                 penalty = "lasso", gamma = 3, maxit = 2500L)

plot(calfit)

calfit.cv <- xval.oem(x = x.matrix, y = medianValue,
                      penalty = "lasso", maxit = 2500L)
                      
plot(calfit.cv)

}
\keyword{datasets}

