% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.oemfit}
\alias{plot.cv.oem}
\alias{plot.oemfit}
\title{Prediction method for Orthogonalizing EM fitted objects}
\usage{
\method{plot}{oemfit}(x, which.model = 1, xvar = c("norm", "lambda",
  "loglambda", "dev"), xlab = iname, ylab = "Coefficients", ...)

\method{plot}{cv.oem}(object, which.model = 1, sign.lambda = 1, ...)
}
\arguments{
\item{x}{fitted "oem" model object}

\item{which.model}{If multiple penalties are fit and returned in the same oem object, the which.model argument is used to 
specify which model to plot. For example, if the oem object "oemobj" was fit with argument 
penalty = c("lasso", "grp.lasso"), then which.model = 2 provides a plot for the group lasso model.}

\item{xvar}{What is on the X-axis. "norm" plots against the L1-norm of the coefficients, "lambda" against the log-lambda sequence, and "dev" 
against the percent deviance explained.}

\item{xlab}{label for x-axis}

\item{ylab}{label for y-axis}

\item{...}{other graphical parameters for the plot}

\item{object}{fitted "cv.oem" model object}

\item{sign.lambda}{Either plot against log(lambda) (default) or its negative if sign.lambda=-1}

\item{which.model}{If multiple penalties are fit and returned in the same oem object, the which.model argument is used to 
specify which model to plot. For example, if the oem object "oemobj" was fit with argument 
penalty = c("lasso", "grp.lasso"), then which.model = 2 provides a plot for the group lasso model.}

\item{...}{other graphical parameters for the plot}
}
\description{
Prediction method for Orthogonalizing EM fitted objects

Prediction method for Orthogonalizing EM fitted objects
}
\examples{
set.seed(123)
n.obs <- 1e4
n.vars <- 100
n.obs.test <- 1e3

true.beta <- c(runif(15, -0.5, 0.5), rep(0, n.vars - 15))

x <- matrix(rnorm(n.obs * n.vars), n.obs, n.vars)
y <- rnorm(n.obs, sd = 3) + x \%*\% true.beta

fit <- oem(x = x, y = y, penalty = c("lasso", "grp.lasso"), groups = rep(1:10, each = 10))

layout(matrix(1:2, ncol = 2))
plot(fit, which.model = 1)
plot(fit, which.model = 2)

set.seed(123)
n.obs <- 1e4
n.vars <- 100
n.obs.test <- 1e3

true.beta <- c(runif(15, -0.5, 0.5), rep(0, n.vars - 15))

x <- matrix(rnorm(n.obs * n.vars), n.obs, n.vars)
y <- rnorm(n.obs, sd = 3) + x \%*\% true.beta

fit <- cv.oem(x = x, y = y, penalty = c("lasso", "grp.lasso"), groups = rep(1:10, each = 10))

layout(matrix(1:2, ncol = 2))
plot(fit, which.model = 1)
plot(fit, which.model = 2)

}

