% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.cv.oem}
\alias{print.summary.cv.oem}
\title{print method for \code{summary.cv.oem} objects}
\usage{
\method{print}{summary.cv.oem}(x, digits, ...)
}
\arguments{
\item{x}{a "summary.cv.oem" object}

\item{digits}{digits to display}

\item{...}{not used}
}
\description{
print method for \code{summary.cv.oem} objects
}
